Network.AWS.StepFunctions

Description

AWS Step Functions

AWS Step Functions is a web service that enables you to coordinate the components of distributed applications and microservices using visual workflows. You build applications from individual components that each perform a discrete function, or task , allowing you to scale and change applications quickly. Step Functions provides a graphical console to visualize the components of your application as a series of steps. It automatically triggers and tracks each step, and retries when there are errors, so your application executes in order and as expected, every time. Step Functions logs the state of each step, so when things do go wrong, you can diagnose and debug problems quickly.

Step Functions manages the operations and underlying infrastructure for you to ensure your application is available at any scale. You can run tasks on the AWS cloud, on your own servers, or an any system that has access to AWS. Step Functions can be accessed and used with the Step Functions console, the AWS SDKs (included with your Beta release invitation email), or an HTTP API (the subject of this document).

Service Configuration

stepFunctions :: Service

API version 2016-11-23 of the Amazon Step Functions SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by StepFunctions.

ExecutionLimitExceeded

_ExecutionLimitExceeded :: AsError a => Getting (First ServiceError) a ServiceError

The maximum number of running executions has been reached. Running executions must end or be stopped before a new execution can be started.

InvalidDefinition

_InvalidDefinition :: AsError a => Getting (First ServiceError) a ServiceError

The provided Amazon States Language definition is invalid.

StateMachineLimitExceeded

_StateMachineLimitExceeded :: AsError a => Getting (First ServiceError) a ServiceError

The maximum number of state machines has been reached. Existing state machines must be deleted before a new state machine can be created.

ExecutionAlreadyExists

_ExecutionAlreadyExists :: AsError a => Getting (First ServiceError) a ServiceError

An execution with the same name already exists.

StateMachineAlreadyExists

_StateMachineAlreadyExists :: AsError a => Getting (First ServiceError) a ServiceError

A state machine with the same name but a different definition or role ARN already exists.

TaskTimedOut

_TaskTimedOut :: AsError a => Getting (First ServiceError) a ServiceError

Prism for TaskTimedOut' errors.

InvalidExecutionInput

_InvalidExecutionInput :: AsError a => Getting (First ServiceError) a ServiceError

The provided JSON input data is invalid.

InvalidOutput

_InvalidOutput :: AsError a => Getting (First ServiceError) a ServiceError

The provided JSON output data is invalid.

InvalidName

_InvalidName :: AsError a => Getting (First ServiceError) a ServiceError

The provided name is invalid.

TaskDoesNotExist

_TaskDoesNotExist :: AsError a => Getting (First ServiceError) a ServiceError

Prism for TaskDoesNotExist' errors.

ActivityDoesNotExist

_ActivityDoesNotExist :: AsError a => Getting (First ServiceError) a ServiceError

The specified activity does not exist.

StateMachineDeleting

_StateMachineDeleting :: AsError a => Getting (First ServiceError) a ServiceError

The specified state machine is being deleted.

InvalidARN

_InvalidARN :: AsError a => Getting (First ServiceError) a ServiceError

The provided Amazon Resource Name (ARN) is invalid.

InvalidToken

_InvalidToken :: AsError a => Getting (First ServiceError) a ServiceError

The provided token is invalid.

ActivityWorkerLimitExceeded

_ActivityWorkerLimitExceeded :: AsError a => Getting (First ServiceError) a ServiceError

The maximum number of workers concurrently polling for activity tasks has been reached.

ActivityLimitExceeded

_ActivityLimitExceeded :: AsError a => Getting (First ServiceError) a ServiceError

The maximum number of activities has been reached. Existing activities must be deleted before a new activity can be created.

ExecutionDoesNotExist

_ExecutionDoesNotExist :: AsError a => Getting (First ServiceError) a ServiceError

The specified execution does not exist.

StateMachineDoesNotExist

_StateMachineDoesNotExist :: AsError a => Getting (First ServiceError) a ServiceError

The specified state machine does not exist.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

DeleteActivity

module Network.AWS.StepFunctions.DeleteActivity

DescribeStateMachine

module Network.AWS.StepFunctions.DescribeStateMachine

StopExecution

module Network.AWS.StepFunctions.StopExecution

DescribeActivity

module Network.AWS.StepFunctions.DescribeActivity

ListStateMachines (Paginated)

module Network.AWS.StepFunctions.ListStateMachines

ListExecutions (Paginated)

module Network.AWS.StepFunctions.ListExecutions

DeleteStateMachine

module Network.AWS.StepFunctions.DeleteStateMachine

GetActivityTask

module Network.AWS.StepFunctions.GetActivityTask

CreateActivity

module Network.AWS.StepFunctions.CreateActivity

ListActivities (Paginated)

module Network.AWS.StepFunctions.ListActivities

SendTaskHeartbeat

module Network.AWS.StepFunctions.SendTaskHeartbeat

SendTaskFailure

module Network.AWS.StepFunctions.SendTaskFailure

DescribeExecution

module Network.AWS.StepFunctions.DescribeExecution

SendTaskSuccess

module Network.AWS.StepFunctions.SendTaskSuccess

StartExecution

module Network.AWS.StepFunctions.StartExecution

GetExecutionHistory (Paginated)

module Network.AWS.StepFunctions.GetExecutionHistory

CreateStateMachine

module Network.AWS.StepFunctions.CreateStateMachine

Types

ExecutionStatus

data ExecutionStatus

Constructors

Aborted

Failed

Running

Succeeded

TimedOut

HistoryEventType

data HistoryEventType

Constructors

ActivityFailed

ActivityScheduleFailed

ActivityScheduled

ActivityStarted

ActivitySucceeded

ActivityTimedOut

ChoiceStateEntered

ChoiceStateExited

ExecutionAborted

ExecutionFailed

ExecutionStarted

ExecutionSucceeded

ExecutionTimedOut

FailStateEntered

LambdaFunctionFailed

LambdaFunctionScheduleFailed

LambdaFunctionScheduled

LambdaFunctionStartFailed

LambdaFunctionStarted

LambdaFunctionSucceeded

LambdaFunctionTimedOut

ParallelStateEntered

ParallelStateExited

PassStateEntered

PassStateExited

SucceedStateEntered

SucceedStateExited

TaskStateEntered

TaskStateExited

WaitStateEntered

WaitStateExited

StateMachineStatus

data StateMachineStatus

Constructors

Active

Deleting

ActivityFailedEventDetails

data ActivityFailedEventDetails

See: activityFailedEventDetails smart constructor.

activityFailedEventDetails :: ActivityFailedEventDetails

Creates a value of ActivityFailedEventDetails with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

afedError - The error code of the failure.

afedCause - A more detailed explanation of the cause of the failure.

afedError :: Lens' ActivityFailedEventDetails (Maybe Text)

The error code of the failure.

afedCause :: Lens' ActivityFailedEventDetails (Maybe Text)

A more detailed explanation of the cause of the failure.

ActivityListItem

data ActivityListItem

See: activityListItem smart constructor.

activityListItem

Arguments

:: Text
aliActivityARN

-> Text
aliName

-> UTCTime
aliCreationDate

-> ActivityListItem

Creates a value of ActivityListItem with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aliActivityARN - The Amazon Resource Name (ARN) that identifies the activity.

aliName - The name of the activity.

aliCreationDate - The date the activity was created.

aliActivityARN :: Lens' ActivityListItem Text

The Amazon Resource Name (ARN) that identifies the activity.

aliName :: Lens' ActivityListItem Text

The name of the activity.

aliCreationDate :: Lens' ActivityListItem UTCTime

The date the activity was created.

ActivityScheduleFailedEventDetails

data ActivityScheduleFailedEventDetails

See: activityScheduleFailedEventDetails smart constructor.

activityScheduleFailedEventDetails :: ActivityScheduleFailedEventDetails

Creates a value of ActivityScheduleFailedEventDetails with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

asfedError - The error code of the failure.

asfedCause - A more detailed explanation of the cause of the failure.

asfedError :: Lens' ActivityScheduleFailedEventDetails (Maybe Text)

The error code of the failure.

asfedCause :: Lens' ActivityScheduleFailedEventDetails (Maybe Text)

A more detailed explanation of the cause of the failure.

ActivityScheduledEventDetails

data ActivityScheduledEventDetails

See: activityScheduledEventDetails smart constructor.

activityScheduledEventDetails

Arguments

:: Text
asedResource

-> ActivityScheduledEventDetails

Creates a value of ActivityScheduledEventDetails with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

asedHeartbeatInSeconds - The maximum allowed duration between two heartbeats for the activity task.

asedInput - The JSON data input to the activity task.

asedTimeoutInSeconds - The maximum allowed duration of the activity task.

asedResource - The Amazon Resource Name (ARN) of the scheduled activity.

asedHeartbeatInSeconds :: Lens' ActivityScheduledEventDetails (Maybe Integer)

The maximum allowed duration between two heartbeats for the activity task.

asedInput :: Lens' ActivityScheduledEventDetails (Maybe Text)

The JSON data input to the activity task.

asedTimeoutInSeconds :: Lens' ActivityScheduledEventDetails (Maybe Integer)

The maximum allowed duration of the activity task.

asedResource :: Lens' ActivityScheduledEventDetails Text

The Amazon Resource Name (ARN) of the scheduled activity.

ActivityStartedEventDetails

data ActivityStartedEventDetails

See: activityStartedEventDetails smart constructor.

activityStartedEventDetails :: ActivityStartedEventDetails

Creates a value of ActivityStartedEventDetails with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

asedWorkerName - The name of the worker that the task was assigned to. These names are provided by the workers when calling GetActivityTask .

asedWorkerName :: Lens' ActivityStartedEventDetails (Maybe Text)

The name of the worker that the task was assigned to. These names are provided by the workers when calling GetActivityTask .

ActivitySucceededEventDetails

data ActivitySucceededEventDetails

See: activitySucceededEventDetails smart constructor.

activitySucceededEventDetails :: ActivitySucceededEventDetails

Creates a value of ActivitySucceededEventDetails with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

asedOutput - The JSON data output by the activity task.

asedOutput :: Lens' ActivitySucceededEventDetails (Maybe Text)

The JSON data output by the activity task.

ActivityTimedOutEventDetails

data ActivityTimedOutEventDetails

See: activityTimedOutEventDetails smart constructor.

activityTimedOutEventDetails :: ActivityTimedOutEventDetails

Creates a value of ActivityTimedOutEventDetails with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

atoedError - The error code of the failure.

atoedCause - A more detailed explanation of the cause of the timeout.

atoedError :: Lens' ActivityTimedOutEventDetails (Maybe Text)

The error code of the failure.

atoedCause :: Lens' ActivityTimedOutEventDetails (Maybe Text)

A more detailed explanation of the cause of the timeout.

ExecutionAbortedEventDetails

data ExecutionAbortedEventDetails

See: executionAbortedEventDetails smart constructor.

executionAbortedEventDetails :: ExecutionAbortedEventDetails

Creates a value of ExecutionAbortedEventDetails with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eaedError - The error code of the failure.

eaedCause - A more detailed explanation of the cause of the failure.

eaedError :: Lens' ExecutionAbortedEventDetails (Maybe Text)

The error code of the failure.

eaedCause :: Lens' ExecutionAbortedEventDetails (Maybe Text)

A more detailed explanation of the cause of the failure.

ExecutionFailedEventDetails

data ExecutionFailedEventDetails

See: executionFailedEventDetails smart constructor.

executionFailedEventDetails :: ExecutionFailedEventDetails

Creates a value of ExecutionFailedEventDetails with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

efedError - The error code of the failure.

efedCause - A more detailed explanation of the cause of the failure.

efedError :: Lens' ExecutionFailedEventDetails (Maybe Text)

The error code of the failure.

efedCause :: Lens' ExecutionFailedEventDetails (Maybe Text)

A more detailed explanation of the cause of the failure.

ExecutionListItem

data ExecutionListItem

See: executionListItem smart constructor.

executionListItem

Arguments

:: Text
eliExecutionARN

-> Text
eliStateMachineARN

-> Text
eliName

-> ExecutionStatus
eliStatus

-> UTCTime
eliStartDate

-> ExecutionListItem

Creates a value of ExecutionListItem with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eliStopDate - If the execution already ended, the date the execution stopped.

eliExecutionARN - The Amazon Resource Name (ARN) that identifies the execution.

eliStateMachineARN - The Amazon Resource Name (ARN) of the executed state machine.

eliName - The name of the execution.

eliStatus - The current status of the execution.

eliStartDate - The date the execution started.

eliStopDate :: Lens' ExecutionListItem (Maybe UTCTime)

If the execution already ended, the date the execution stopped.

eliExecutionARN :: Lens' ExecutionListItem Text

The Amazon Resource Name (ARN) that identifies the execution.

eliStateMachineARN :: Lens' ExecutionListItem Text

The Amazon Resource Name (ARN) of the executed state machine.

eliName :: Lens' ExecutionListItem Text

The name of the execution.

eliStatus :: Lens' ExecutionListItem ExecutionStatus

The current status of the execution.

eliStartDate :: Lens' ExecutionListItem UTCTime

The date the execution started.

ExecutionStartedEventDetails

data ExecutionStartedEventDetails

See: executionStartedEventDetails smart constructor.

executionStartedEventDetails :: ExecutionStartedEventDetails

Creates a value of ExecutionStartedEventDetails with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

esedInput - The JSON data input to the execution.

esedRoleARN - The Amazon Resource Name (ARN) of the IAM role used for executing AWS Lambda tasks.

esedInput :: Lens' ExecutionStartedEventDetails (Maybe Text)

The JSON data input to the execution.

esedRoleARN :: Lens' ExecutionStartedEventDetails (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role used for executing AWS Lambda tasks.

ExecutionSucceededEventDetails

data ExecutionSucceededEventDetails

See: executionSucceededEventDetails smart constructor.

executionSucceededEventDetails :: ExecutionSucceededEventDetails

Creates a value of ExecutionSucceededEventDetails with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

esedOutput - The JSON data output by the execution.

esedOutput :: Lens' ExecutionSucceededEventDetails (Maybe Text)

The JSON data output by the execution.

ExecutionTimedOutEventDetails

data ExecutionTimedOutEventDetails

See: executionTimedOutEventDetails smart constructor.

executionTimedOutEventDetails :: ExecutionTimedOutEventDetails

Creates a value of ExecutionTimedOutEventDetails with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

etoedError - The error code of the failure.

etoedCause - A more detailed explanation of the cause of the timeout.

etoedError :: Lens' ExecutionTimedOutEventDetails (Maybe Text)

The error code of the failure.

etoedCause :: Lens' ExecutionTimedOutEventDetails (Maybe Text)

A more detailed explanation of the cause of the timeout.

HistoryEvent

data HistoryEvent

See: historyEvent smart constructor.

historyEvent

Arguments

:: UTCTime
heTimestamp

-> HistoryEventType
heType

-> Integer
heId

-> HistoryEvent

Creates a value of HistoryEvent with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

heActivityStartedEventDetails - Undocumented member.

heLambdaFunctionStartFailedEventDetails - Undocumented member.

heStateExitedEventDetails - Undocumented member.

heLambdaFunctionSucceededEventDetails - Undocumented member.

heActivitySucceededEventDetails - Undocumented member.

heLambdaFunctionTimedOutEventDetails - Undocumented member.

heActivityTimedOutEventDetails - Undocumented member.

heExecutionFailedEventDetails - Undocumented member.

heExecutionAbortedEventDetails - Undocumented member.

heExecutionSucceededEventDetails - Undocumented member.

heLambdaFunctionScheduledEventDetails - Undocumented member.

heActivityScheduledEventDetails - Undocumented member.

heExecutionStartedEventDetails - Undocumented member.

heActivityScheduleFailedEventDetails - Undocumented member.

heLambdaFunctionScheduleFailedEventDetails - Undocumented member.

heStateEnteredEventDetails - Undocumented member.

hePreviousEventId - The id of the previous event.

heActivityFailedEventDetails - Undocumented member.

heLambdaFunctionFailedEventDetails - Undocumented member.

heExecutionTimedOutEventDetails - Undocumented member.

heTimestamp - The date the event occured.

heType - The type of the event.

heId - The id of the event. Events are numbered sequentially, starting at one.

heActivityStartedEventDetails :: Lens' HistoryEvent (Maybe ActivityStartedEventDetails)

Undocumented member.

heLambdaFunctionStartFailedEventDetails :: Lens' HistoryEvent (Maybe LambdaFunctionStartFailedEventDetails)

Undocumented member.

heStateExitedEventDetails :: Lens' HistoryEvent (Maybe StateExitedEventDetails)

Undocumented member.

heLambdaFunctionSucceededEventDetails :: Lens' HistoryEvent (Maybe LambdaFunctionSucceededEventDetails)

Undocumented member.

heActivitySucceededEventDetails :: Lens' HistoryEvent (Maybe ActivitySucceededEventDetails)

Undocumented member.

heLambdaFunctionTimedOutEventDetails :: Lens' HistoryEvent (Maybe LambdaFunctionTimedOutEventDetails)

Undocumented member.

heActivityTimedOutEventDetails :: Lens' HistoryEvent (Maybe ActivityTimedOutEventDetails)

Undocumented member.

heExecutionFailedEventDetails :: Lens' HistoryEvent (Maybe ExecutionFailedEventDetails)

Undocumented member.

heExecutionAbortedEventDetails :: Lens' HistoryEvent (Maybe ExecutionAbortedEventDetails)

Undocumented member.

heExecutionSucceededEventDetails :: Lens' HistoryEvent (Maybe ExecutionSucceededEventDetails)

Undocumented member.

heLambdaFunctionScheduledEventDetails :: Lens' HistoryEvent (Maybe LambdaFunctionScheduledEventDetails)

Undocumented member.

heActivityScheduledEventDetails :: Lens' HistoryEvent (Maybe ActivityScheduledEventDetails)

Undocumented member.

heExecutionStartedEventDetails :: Lens' HistoryEvent (Maybe ExecutionStartedEventDetails)

Undocumented member.

heActivityScheduleFailedEventDetails :: Lens' HistoryEvent (Maybe ActivityScheduleFailedEventDetails)

Undocumented member.

heLambdaFunctionScheduleFailedEventDetails :: Lens' HistoryEvent (Maybe LambdaFunctionScheduleFailedEventDetails)

Undocumented member.

heStateEnteredEventDetails :: Lens' HistoryEvent (Maybe StateEnteredEventDetails)

Undocumented member.

hePreviousEventId :: Lens' HistoryEvent (Maybe Integer)

The id of the previous event.

heActivityFailedEventDetails :: Lens' HistoryEvent (Maybe ActivityFailedEventDetails)

Undocumented member.

heLambdaFunctionFailedEventDetails :: Lens' HistoryEvent (Maybe LambdaFunctionFailedEventDetails)

Undocumented member.

heExecutionTimedOutEventDetails :: Lens' HistoryEvent (Maybe ExecutionTimedOutEventDetails)

Undocumented member.

heTimestamp :: Lens' HistoryEvent UTCTime

The date the event occured.

heType :: Lens' HistoryEvent HistoryEventType

The type of the event.

heId :: Lens' HistoryEvent Integer

The id of the event. Events are numbered sequentially, starting at one.

LambdaFunctionFailedEventDetails

data LambdaFunctionFailedEventDetails

See: lambdaFunctionFailedEventDetails smart constructor.

lambdaFunctionFailedEventDetails :: LambdaFunctionFailedEventDetails

Creates a value of LambdaFunctionFailedEventDetails with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lffedError - The error code of the failure.

lffedCause - A more detailed explanation of the cause of the failure.

lffedError :: Lens' LambdaFunctionFailedEventDetails (Maybe Text)

The error code of the failure.

lffedCause :: Lens' LambdaFunctionFailedEventDetails (Maybe Text)

A more detailed explanation of the cause of the failure.

LambdaFunctionScheduleFailedEventDetails

data LambdaFunctionScheduleFailedEventDetails

See: lambdaFunctionScheduleFailedEventDetails smart constructor.

lambdaFunctionScheduleFailedEventDetails :: LambdaFunctionScheduleFailedEventDetails

Creates a value of LambdaFunctionScheduleFailedEventDetails with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lError - The error code of the failure.

lCause - A more detailed explanation of the cause of the failure.

lError :: Lens' LambdaFunctionScheduleFailedEventDetails (Maybe Text)

The error code of the failure.

lCause :: Lens' LambdaFunctionScheduleFailedEventDetails (Maybe Text)

A more detailed explanation of the cause of the failure.

LambdaFunctionScheduledEventDetails

data LambdaFunctionScheduledEventDetails

See: lambdaFunctionScheduledEventDetails smart constructor.

lambdaFunctionScheduledEventDetails

Arguments

:: Text
lfsedResource

-> LambdaFunctionScheduledEventDetails

Creates a value of LambdaFunctionScheduledEventDetails with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lfsedInput - The JSON data input to the lambda function.

lfsedTimeoutInSeconds - The maximum allowed duration of the lambda function.

lfsedResource - The Amazon Resource Name (ARN) of the scheduled lambda function.

lfsedInput :: Lens' LambdaFunctionScheduledEventDetails (Maybe Text)

The JSON data input to the lambda function.

lfsedTimeoutInSeconds :: Lens' LambdaFunctionScheduledEventDetails (Maybe Integer)

The maximum allowed duration of the lambda function.

lfsedResource :: Lens' LambdaFunctionScheduledEventDetails Text

The Amazon Resource Name (ARN) of the scheduled lambda function.

LambdaFunctionStartFailedEventDetails

data LambdaFunctionStartFailedEventDetails

See: lambdaFunctionStartFailedEventDetails smart constructor.

lambdaFunctionStartFailedEventDetails :: LambdaFunctionStartFailedEventDetails

Creates a value of LambdaFunctionStartFailedEventDetails with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lfsfedError - The error code of the failure.

lfsfedCause - A more detailed explanation of the cause of the failure.

lfsfedError :: Lens' LambdaFunctionStartFailedEventDetails (Maybe Text)

The error code of the failure.

lfsfedCause :: Lens' LambdaFunctionStartFailedEventDetails (Maybe Text)

A more detailed explanation of the cause of the failure.

LambdaFunctionSucceededEventDetails

data LambdaFunctionSucceededEventDetails

See: lambdaFunctionSucceededEventDetails smart constructor.

lambdaFunctionSucceededEventDetails :: LambdaFunctionSucceededEventDetails

Creates a value of LambdaFunctionSucceededEventDetails with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lfsedOutput - The JSON data output by the lambda function.

lfsedOutput :: Lens' LambdaFunctionSucceededEventDetails (Maybe Text)

The JSON data output by the lambda function.

LambdaFunctionTimedOutEventDetails

data LambdaFunctionTimedOutEventDetails

See: lambdaFunctionTimedOutEventDetails smart constructor.

lambdaFunctionTimedOutEventDetails :: LambdaFunctionTimedOutEventDetails

Creates a value of LambdaFunctionTimedOutEventDetails with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lftoedError - The error code of the failure.

lftoedCause - A more detailed explanation of the cause of the timeout.

lftoedError :: Lens' LambdaFunctionTimedOutEventDetails (Maybe Text)

The error code of the failure.

lftoedCause :: Lens' LambdaFunctionTimedOutEventDetails (Maybe Text)

A more detailed explanation of the cause of the timeout.

StateEnteredEventDetails

data StateEnteredEventDetails

See: stateEnteredEventDetails smart constructor.

stateEnteredEventDetails

Arguments

:: Text
sName

-> StateEnteredEventDetails

Creates a value of StateEnteredEventDetails with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sInput - The JSON input data to the state.

sName - The name of the state.

sInput :: Lens' StateEnteredEventDetails (Maybe Text)

The JSON input data to the state.

sName :: Lens' StateEnteredEventDetails Text

The name of the state.

StateExitedEventDetails

data StateExitedEventDetails

See: stateExitedEventDetails smart constructor.

stateExitedEventDetails

Arguments

:: Text
seedName

-> StateExitedEventDetails

Creates a value of StateExitedEventDetails with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

seedOutput - The JSON output data of the state.

seedName - The name of the state.

seedOutput :: Lens' StateExitedEventDetails (Maybe Text)

The JSON output data of the state.

seedName :: Lens' StateExitedEventDetails Text

The name of the state.

StateMachineListItem

data StateMachineListItem

See: stateMachineListItem smart constructor.

stateMachineListItem

Arguments

:: Text
smliStateMachineARN

-> Text
smliName

-> UTCTime
smliCreationDate

-> StateMachineListItem

Creates a value of StateMachineListItem with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

smliStateMachineARN - The Amazon Resource Name (ARN) that identifies the state machine.

smliName - The name of the state machine.

smliCreationDate - The date the state machine was created.

smliStateMachineARN :: Lens' StateMachineListItem Text

The Amazon Resource Name (ARN) that identifies the state machine.

smliName :: Lens' StateMachineListItem Text

The name of the state machine.

smliCreationDate :: Lens' StateMachineListItem UTCTime

The date the state machine was created.