Network.AWS.StepFunctions.CreateActivity

Description

Creates an activity.

Creating a Request

createActivity

Arguments

:: Text
caName

-> CreateActivity

Creates a value of CreateActivity with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

caName - The name of the activity to create. This name must be unique for your AWS account and region.

data CreateActivity

See: createActivity smart constructor.

Request Lenses

caName :: Lens' CreateActivity Text

The name of the activity to create. This name must be unique for your AWS account and region.

Destructuring the Response

createActivityResponse

Arguments

:: Int
carsResponseStatus

-> Text
carsActivityARN

-> UTCTime
carsCreationDate

-> CreateActivityResponse

Creates a value of CreateActivityResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

carsResponseStatus - -- | The response status code.

carsActivityARN - The Amazon Resource Name (ARN) that identifies the created activity.

carsCreationDate - The date the activity was created.

data CreateActivityResponse

See: createActivityResponse smart constructor.

Response Lenses

carsResponseStatus :: Lens' CreateActivityResponse Int

- | The response status code.

carsActivityARN :: Lens' CreateActivityResponse Text

The Amazon Resource Name (ARN) that identifies the created activity.

carsCreationDate :: Lens' CreateActivityResponse UTCTime

The date the activity was created.