Network.AWS.StepFunctions.DescribeActivity

Description

Describes an activity.

Creating a Request

describeActivity

Arguments

:: Text
dActivityARN

-> DescribeActivity

Creates a value of DescribeActivity with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dActivityARN - The Amazon Resource Name (ARN) of the activity to describe.

data DescribeActivity

See: describeActivity smart constructor.

Request Lenses

dActivityARN :: Lens' DescribeActivity Text

The Amazon Resource Name (ARN) of the activity to describe.

Destructuring the Response

describeActivityResponse

Arguments

:: Int
desrsResponseStatus

-> Text
desrsActivityARN

-> Text
desrsName

-> UTCTime
desrsCreationDate

-> DescribeActivityResponse

Creates a value of DescribeActivityResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

desrsResponseStatus - -- | The response status code.

desrsActivityARN - The Amazon Resource Name (ARN) that identifies the activity.

desrsName - The name of the activity.

desrsCreationDate - The date the activity was created.

data DescribeActivityResponse

See: describeActivityResponse smart constructor.

Response Lenses

desrsResponseStatus :: Lens' DescribeActivityResponse Int

- | The response status code.

desrsActivityARN :: Lens' DescribeActivityResponse Text

The Amazon Resource Name (ARN) that identifies the activity.

desrsName :: Lens' DescribeActivityResponse Text

The name of the activity.

desrsCreationDate :: Lens' DescribeActivityResponse UTCTime

The date the activity was created.