Network.AWS.StepFunctions.DescribeExecution

Description

Describes an execution.

Creating a Request

describeExecution

Arguments

:: Text
deExecutionARN

-> DescribeExecution

Creates a value of DescribeExecution with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

deExecutionARN - The Amazon Resource Name (ARN) of the execution to describe.

data DescribeExecution

See: describeExecution smart constructor.

Request Lenses

deExecutionARN :: Lens' DescribeExecution Text

The Amazon Resource Name (ARN) of the execution to describe.

Destructuring the Response

describeExecutionResponse

Arguments

:: Int
dersResponseStatus

-> Text
dersExecutionARN

-> Text
dersStateMachineARN

-> ExecutionStatus
dersStatus

-> UTCTime
dersStartDate

-> Text
dersInput

-> DescribeExecutionResponse

Creates a value of DescribeExecutionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dersStopDate - If the execution has already ended, the date the execution stopped.

dersName - The name of the execution.

dersOutput - The JSON output data of the execution.

dersResponseStatus - -- | The response status code.

dersExecutionARN - The Amazon Resource Name (ARN) that identifies the execution.

dersStateMachineARN - The Amazon Resource Name (ARN) of the executed stated machine.

dersStatus - The current status of the execution.

dersStartDate - The date the execution was started.

dersInput - The JSON input data of the execution.

data DescribeExecutionResponse

See: describeExecutionResponse smart constructor.

Response Lenses

dersStopDate :: Lens' DescribeExecutionResponse (Maybe UTCTime)

If the execution has already ended, the date the execution stopped.

dersName :: Lens' DescribeExecutionResponse (Maybe Text)

The name of the execution.

dersOutput :: Lens' DescribeExecutionResponse (Maybe Text)

The JSON output data of the execution.

dersResponseStatus :: Lens' DescribeExecutionResponse Int

- | The response status code.

dersExecutionARN :: Lens' DescribeExecutionResponse Text

The Amazon Resource Name (ARN) that identifies the execution.

dersStateMachineARN :: Lens' DescribeExecutionResponse Text

The Amazon Resource Name (ARN) of the executed stated machine.

dersStatus :: Lens' DescribeExecutionResponse ExecutionStatus

The current status of the execution.

dersStartDate :: Lens' DescribeExecutionResponse UTCTime

The date the execution was started.

dersInput :: Lens' DescribeExecutionResponse Text

The JSON input data of the execution.