Network.AWS.StepFunctions.DescribeStateMachine

Description

Describes a state machine.

Creating a Request

describeStateMachine

Arguments

:: Text
dsmStateMachineARN

-> DescribeStateMachine

Creates a value of DescribeStateMachine with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsmStateMachineARN - The Amazon Resource Name (ARN) of the state machine to describe.

data DescribeStateMachine

See: describeStateMachine smart constructor.

Request Lenses

dsmStateMachineARN :: Lens' DescribeStateMachine Text

The Amazon Resource Name (ARN) of the state machine to describe.

Destructuring the Response

describeStateMachineResponse

Arguments

:: Int
dsmrsResponseStatus

-> Text
dsmrsStateMachineARN

-> Text
dsmrsName

-> Text
dsmrsDefinition

-> Text
dsmrsRoleARN

-> UTCTime
dsmrsCreationDate

-> DescribeStateMachineResponse

Creates a value of DescribeStateMachineResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsmrsStatus - The current status of the state machine.

dsmrsResponseStatus - -- | The response status code.

dsmrsStateMachineARN - The Amazon Resource Name (ARN) that identifies the state machine.

dsmrsName - The name of the state machine.

dsmrsDefinition - The Amazon States Language definition of the state machine.

dsmrsRoleARN - The Amazon Resource Name (ARN) of the IAM role used for executing this state machine.

dsmrsCreationDate - The date the state machine was created.

data DescribeStateMachineResponse

See: describeStateMachineResponse smart constructor.

Response Lenses

dsmrsStatus :: Lens' DescribeStateMachineResponse (Maybe StateMachineStatus)

The current status of the state machine.

dsmrsResponseStatus :: Lens' DescribeStateMachineResponse Int

- | The response status code.

dsmrsStateMachineARN :: Lens' DescribeStateMachineResponse Text

The Amazon Resource Name (ARN) that identifies the state machine.

dsmrsName :: Lens' DescribeStateMachineResponse Text

The name of the state machine.

dsmrsDefinition :: Lens' DescribeStateMachineResponse Text

The Amazon States Language definition of the state machine.

dsmrsRoleARN :: Lens' DescribeStateMachineResponse Text

The Amazon Resource Name (ARN) of the IAM role used for executing this state machine.

dsmrsCreationDate :: Lens' DescribeStateMachineResponse UTCTime

The date the state machine was created.