Network.AWS.StepFunctions.GetActivityTask

Description

Used by workers to retrieve a task (with the specified activity ARN) scheduled for execution by a running state machine. This initiates a long poll, where the service holds the HTTP connection open and responds as soon as a task becomes available (i.e. an execution of a task of this type is needed.) The maximum time the service holds on to the request before responding is 60 seconds. If no task is available within 60 seconds, the poll will return an empty result, that is, the taskToken returned is an empty string.

Important: Workers should set their client side socket timeout to at least 65 seconds (5 seconds higher than the maximum time the service may hold the poll request).

Creating a Request

getActivityTask

Arguments

:: Text
gatActivityARN

-> GetActivityTask

Creates a value of GetActivityTask with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gatWorkerName - An arbitrary name may be provided in order to identify the worker that the task is assigned to. This name will be used when it is logged in the execution history.

gatActivityARN - The Amazon Resource Name (ARN) of the activity to retrieve tasks from.

data GetActivityTask

See: getActivityTask smart constructor.

Request Lenses

gatWorkerName :: Lens' GetActivityTask (Maybe Text)

An arbitrary name may be provided in order to identify the worker that the task is assigned to. This name will be used when it is logged in the execution history.

gatActivityARN :: Lens' GetActivityTask Text

The Amazon Resource Name (ARN) of the activity to retrieve tasks from.

Destructuring the Response

getActivityTaskResponse

Arguments

:: Int
gatrsResponseStatus

-> GetActivityTaskResponse

Creates a value of GetActivityTaskResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gatrsInput - The JSON input data for the task.

gatrsTaskToken - A token that identifies the scheduled task. This token must be copied and included in subsequent calls to SendTaskHeartbeat , SendTaskSuccess or SendTaskFailure in order to report the progress or completion of the task.

gatrsResponseStatus - -- | The response status code.

data GetActivityTaskResponse

See: getActivityTaskResponse smart constructor.

Response Lenses

gatrsInput :: Lens' GetActivityTaskResponse (Maybe Text)

The JSON input data for the task.

gatrsTaskToken :: Lens' GetActivityTaskResponse (Maybe Text)

A token that identifies the scheduled task. This token must be copied and included in subsequent calls to SendTaskHeartbeat , SendTaskSuccess or SendTaskFailure in order to report the progress or completion of the task.

gatrsResponseStatus :: Lens' GetActivityTaskResponse Int

- | The response status code.