Network.AWS.StepFunctions.ListActivities

Description

Lists the existing activities. The results may be split into multiple pages. To retrieve subsequent pages, make the call again using the nextToken returned by the previous call.

This operation returns paginated results.

Creating a Request

listActivities :: ListActivities

Creates a value of ListActivities with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

laNextToken - If a nextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextToken . Keep all other arguments unchanged. The configured maxResults determines how many results can be returned in a single call.

laMaxResults - The maximum number of results that will be returned per call. nextToken can be used to obtain further pages of results. The default is 100 and the maximum allowed page size is 1000. This is an upper limit only; the actual number of results returned per call may be fewer than the specified maximum.

data ListActivities

See: listActivities smart constructor.

Request Lenses

laNextToken :: Lens' ListActivities (Maybe Text)

If a nextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextToken . Keep all other arguments unchanged. The configured maxResults determines how many results can be returned in a single call.

laMaxResults :: Lens' ListActivities (Maybe Natural)

The maximum number of results that will be returned per call. nextToken can be used to obtain further pages of results. The default is 100 and the maximum allowed page size is 1000. This is an upper limit only; the actual number of results returned per call may be fewer than the specified maximum.

Destructuring the Response

listActivitiesResponse

Arguments

:: Int
larsResponseStatus

-> ListActivitiesResponse

Creates a value of ListActivitiesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

larsNextToken - If a nextToken is returned, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextToken . Keep all other arguments unchanged. The configured maxResults determines how many results can be returned in a single call.

larsResponseStatus - -- | The response status code.

larsActivities - The list of activities.

data ListActivitiesResponse

See: listActivitiesResponse smart constructor.

Response Lenses

larsNextToken :: Lens' ListActivitiesResponse (Maybe Text)

If a nextToken is returned, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextToken . Keep all other arguments unchanged. The configured maxResults determines how many results can be returned in a single call.

larsResponseStatus :: Lens' ListActivitiesResponse Int

- | The response status code.

larsActivities :: Lens' ListActivitiesResponse [ActivityListItem]

The list of activities.