Network.AWS.StepFunctions.ListStateMachines

Description

Lists the existing state machines. The results may be split into multiple pages. To retrieve subsequent pages, make the call again using the nextToken returned by the previous call.

This operation returns paginated results.

Creating a Request

listStateMachines :: ListStateMachines

Creates a value of ListStateMachines with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lsmNextToken - If a nextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextToken . Keep all other arguments unchanged. The configured maxResults determines how many results can be returned in a single call.

lsmMaxResults - The maximum number of results that will be returned per call. nextToken can be used to obtain further pages of results. The default is 100 and the maximum allowed page size is 1000. This is an upper limit only; the actual number of results returned per call may be fewer than the specified maximum.

data ListStateMachines

See: listStateMachines smart constructor.

Request Lenses

lsmNextToken :: Lens' ListStateMachines (Maybe Text)

If a nextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextToken . Keep all other arguments unchanged. The configured maxResults determines how many results can be returned in a single call.

lsmMaxResults :: Lens' ListStateMachines (Maybe Natural)

The maximum number of results that will be returned per call. nextToken can be used to obtain further pages of results. The default is 100 and the maximum allowed page size is 1000. This is an upper limit only; the actual number of results returned per call may be fewer than the specified maximum.

Destructuring the Response

listStateMachinesResponse

Arguments

:: Int
lsmrsResponseStatus

-> ListStateMachinesResponse

Creates a value of ListStateMachinesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lsmrsNextToken - If a nextToken is returned, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextToken . Keep all other arguments unchanged. The configured maxResults determines how many results can be returned in a single call.

lsmrsResponseStatus - -- | The response status code.

lsmrsStateMachines - Undocumented member.

data ListStateMachinesResponse

See: listStateMachinesResponse smart constructor.

Response Lenses

lsmrsNextToken :: Lens' ListStateMachinesResponse (Maybe Text)

If a nextToken is returned, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextToken . Keep all other arguments unchanged. The configured maxResults determines how many results can be returned in a single call.

lsmrsResponseStatus :: Lens' ListStateMachinesResponse Int

- | The response status code.

lsmrsStateMachines :: Lens' ListStateMachinesResponse [StateMachineListItem]

Undocumented member.