Network.AWS.StepFunctions.SendTaskFailure

Description

Used by workers to report that the task identified by the taskToken failed.

Creating a Request

sendTaskFailure

Arguments

:: Text
stfTaskToken

-> SendTaskFailure

Creates a value of SendTaskFailure with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

stfError - An arbitrary error code that identifies the cause of the failure.

stfCause - A more detailed explanation of the cause of the failure.

stfTaskToken - The token that represents this task. Task tokens are generated by the service when the tasks are assigned to a worker (see GetActivityTask::taskToken).

data SendTaskFailure

See: sendTaskFailure smart constructor.

Request Lenses

stfError :: Lens' SendTaskFailure (Maybe Text)

An arbitrary error code that identifies the cause of the failure.

stfCause :: Lens' SendTaskFailure (Maybe Text)

A more detailed explanation of the cause of the failure.

stfTaskToken :: Lens' SendTaskFailure Text

The token that represents this task. Task tokens are generated by the service when the tasks are assigned to a worker (see GetActivityTask::taskToken).

Destructuring the Response

sendTaskFailureResponse

Arguments

:: Int
stfrsResponseStatus

-> SendTaskFailureResponse

Creates a value of SendTaskFailureResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

stfrsResponseStatus - -- | The response status code.

data SendTaskFailureResponse

See: sendTaskFailureResponse smart constructor.

Response Lenses

stfrsResponseStatus :: Lens' SendTaskFailureResponse Int

- | The response status code.