Network.AWS.StepFunctions.SendTaskSuccess

Description

Used by workers to report that the task identified by the taskToken completed successfully.

Creating a Request

sendTaskSuccess

Arguments

:: Text
stsTaskToken

-> Text
stsOutput

-> SendTaskSuccess

Creates a value of SendTaskSuccess with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

stsTaskToken - The token that represents this task. Task tokens are generated by the service when the tasks are assigned to a worker (see GetActivityTask::taskToken).

stsOutput - The JSON output of the task.

data SendTaskSuccess

See: sendTaskSuccess smart constructor.

Request Lenses

stsTaskToken :: Lens' SendTaskSuccess Text

The token that represents this task. Task tokens are generated by the service when the tasks are assigned to a worker (see GetActivityTask::taskToken).

stsOutput :: Lens' SendTaskSuccess Text

The JSON output of the task.

Destructuring the Response

sendTaskSuccessResponse

Arguments

:: Int
stsrsResponseStatus

-> SendTaskSuccessResponse

Creates a value of SendTaskSuccessResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

stsrsResponseStatus - -- | The response status code.

data SendTaskSuccessResponse

See: sendTaskSuccessResponse smart constructor.

Response Lenses

stsrsResponseStatus :: Lens' SendTaskSuccessResponse Int

- | The response status code.